moving common code to common location
