+++ packages/service-library/src/servicelib/aiohttp/long_running_tasks/__init__.py
@@ -1,0 +1,3 @@
+from . import server
+
+__all__: tuple[str, ...] = ("server",)
