+++ packages/service-library/src/servicelib/fastapi/long_running_tasks/_server.py
@@ -3,10 +3,10 @@ from typing import Final
 from fastapi import APIRouter, FastAPI
 from pydantic import PositiveFloat
 
+from ...long_running_tasks._errors import BaseLongRunningError
+from ...long_running_tasks._task import TaskManager
 from ._error_handlers import base_long_running_error_handler
-from ._errors import BaseLongRunningError
 from ._routes import router
-from ._task import TaskManager
 
 _MINUTE: Final[PositiveFloat] = 60
