+++ packages/service-library/tests/fastapi/test_long_running_tasks_task.py
@@ -11,12 +11,6 @@ from typing import AsyncIterable, AsyncIterator, Optional
 import pytest
 from asgi_lifespan import LifespanManager
 from fastapi import APIRouter, Depends, FastAPI, status
-from servicelib.fastapi.long_running_tasks._errors import (
-    TaskAlreadyRunningError,
-    TaskNotCompletedError,
-    TaskNotFoundError,
-)
-from servicelib.fastapi.long_running_tasks._models import TaskResult
 from servicelib.fastapi.long_running_tasks.server import (
     TaskId,
     TaskManager,
@@ -26,6 +20,12 @@ from servicelib.fastapi.long_running_tasks.server import (
 )
 from servicelib.fastapi.long_running_tasks.server import setup as setup_server
 from servicelib.fastapi.long_running_tasks.server import start_task
+from servicelib.long_running_tasks._errors import (
+    TaskAlreadyRunningError,
+    TaskNotCompletedError,
+    TaskNotFoundError,
+)
+from servicelib.long_running_tasks._models import TaskResult
 
 # UTILS
